Fix less strict code

+++ src/gameover_screen.c
@@ -22,15 +22,15 @@ void game_over_init(GameState* game_state)
     state = game_state;
     Font* font = assetmgr_get_font("main");
 
-    u16 text_height = GAME_OVER_N_ITEMS * UI_MENU_ITEM_SIZE + UI_HEADER_SIZE;
-    u16 starty = GetScreenHeight() * 0.5f - text_height * 0.5f + UI_HEADER_SIZE;
+    u16 text_height = (u16)(GAME_OVER_N_ITEMS * UI_MENU_ITEM_SIZE + UI_HEADER_SIZE);
+    u16 starty = (u16)((f32)GetScreenHeight() * 0.5f - text_height * 0.5f + UI_HEADER_SIZE);
 
     go_mitems[n_go_mitems++] = create_menu_item("Replay Game", 0.0f, starty, font, replay_fn, ALIGN_CENTRE);
 
-    starty += UI_MENU_ITEM_SIZE;
+    starty += (u16)UI_MENU_ITEM_SIZE;
     go_mitems[n_go_mitems++] = create_menu_item("Main Menu", 0.0f, starty, font, main_menu_fn, ALIGN_CENTRE);
 
-    starty += UI_MENU_ITEM_SIZE;
+    starty += (u16)UI_MENU_ITEM_SIZE;
     go_mitems[n_go_mitems++] = create_menu_item("Quit", 0.0f, starty, font, quit_fn, ALIGN_CENTRE);
 }
 
@@ -70,11 +70,11 @@ void game_over_render(void)
 
         // Not affected by camera
         {
-            u16 text_height = n_go_mitems * UI_MENU_ITEM_SIZE + UI_HEADER_SIZE;
-            u16 starty = GetScreenHeight() * 0.5f - text_height * 0.5f;
+            u16 text_height = (u16)((f32)n_go_mitems * UI_MENU_ITEM_SIZE + UI_HEADER_SIZE);
+            u16 starty = (u16)((f32)GetScreenHeight() * 0.5f - text_height * 0.5f);
 
             Vector2 header_size = MeasureTextEx(*font, "Game Over", UI_HEADER_SIZE, 1.0f);
-            u16 headerx = GetScreenWidth() * 0.5f - header_size.x * 0.5f;
+            u16 headerx = (u16)((f32)GetScreenWidth() * 0.5f - header_size.x * 0.5f);
             DrawTextEx(*font, "Game Over", (Vector2){headerx, starty}, UI_HEADER_SIZE, 1.0f, PALEBLUE_D);
 
             for (size_t i = 0; i < n_go_mitems; ++i) {
